Added csView tool to extract source code from svg description field
